src/python/requirements.txt
deepgram-sdk>=2.0.0
python-dotenv>=0.19.0
aiofiles>=23.2.1
